Fix biters not inheriting default ai settings, causing lag
This global was removed.

The settings armoured biters are missing are thus:
{ destroy_when_commands_fail = true,
  allow_try_return_to_spawner = true }

allow_try_return_to_spawner, specifically, causes biters that are not
able to attack to try to return to their spawner, or to despawn as a
failsafe. Without this, large deathworld games pile up active armoured
biters not going anywhere, slowly accumulating massive amounts of lag.

destroy_when_commands_fail is another failsafe for stuck units, so
also bad to be missing.

This reduced the active Units in our SE Death World save from 8500 to
900 (10ms -> 1ms)